requirements.txt
aiohttp==3.10.11
aioari==0.6.4
python-dotenv==1.0.1
requests==2.32.3
